fix: audition (#84)
* Unify runtime/realms data layouts with glam vectors and matrices

* clear

* refactor(material): unify 2d/3d compilation flow without bypass

* refactor(render): remove direct preset material bootstrap from render defaults

* feat(material): move core material ids to reserved range

* test(runtime): fix material builtin-id behavior and active fields in fixtures

* docs(material): document unified realm flow and reserved core ids

* refactor(render): finalize 2d material flow and multi-light shadow integration

* fix(demo): update RUN_DURATION_MS to 8 seconds in demo files

+++ crates/galfus-demo/src/demo/session.rs
@@ -38,7 +38,7 @@ fn wait_for_window_ready(window_id: u32, title: &str) -> WindowBinding {
 
     while start.elapsed() < WINDOW_READY_TIMEOUT {
         assert_eq!(
-            core::galfus_tick(total_ms, WINDOW_READY_DELTA_MS),
+            core::galfus_tick(total_ms as i64, WINDOW_READY_DELTA_MS),
             GalfusResult::Success
         );
         total_ms = total_ms.saturating_add(WINDOW_READY_DELTA_MS as u64);
